+++ routes/api.js
@@ -3,25 +3,25 @@ const ApiController = require('../controllers/apiController');
 
 router.get('/', ApiController.index);
 
-router.use('/ability-scores', require('./abilityScores'));
-router.use('/classes', require('./classes'));
-router.use('/conditions', require('./conditions'));
-router.use('/damage-types', require('./damageTypes'));
-router.use('/equipment-categories', require('./equipmentCategories'));
-router.use('/equipment', require('./equipment'));
-router.use('/features', require('./features'));
-router.use('/languages', require('./languages'));
-router.use('/magic-schools', require('./magicSchools'));
-router.use('/monsters', require('./monsters'));
-router.use('/proficiencies', require('./proficiencies'));
-router.use('/races', require('./races'));
-router.use('/skills', require('./skills'));
-router.use('/spellcasting', require('./spellcasting'));
-router.use('/spells', require('./spells'));
-router.use('/startingequipment', require('./startingEquipment'));
-router.use('/subclasses', require('./subclasses'));
-router.use('/subraces', require('./subraces'));
-router.use('/traits', require('./traits'));
-router.use('/weapon-properties', require('./weaponProperties'));
+router.use('/ability-scores', require('./api/abilityScores'));
+router.use('/classes', require('./api/classes'));
+router.use('/conditions', require('./api/conditions'));
+router.use('/damage-types', require('./api/damageTypes'));
+router.use('/equipment-categories', require('./api/equipmentCategories'));
+router.use('/equipment', require('./api/equipment'));
+router.use('/features', require('./api/features'));
+router.use('/languages', require('./api/languages'));
+router.use('/magic-schools', require('./api/magicSchools'));
+router.use('/monsters', require('./api/monsters'));
+router.use('/proficiencies', require('./api/proficiencies'));
+router.use('/races', require('./api/races'));
+router.use('/skills', require('./api/skills'));
+router.use('/spellcasting', require('./api/spellcasting'));
+router.use('/spells', require('./api/spells'));
+router.use('/startingequipment', require('./api/startingequipment'));
+router.use('/subclasses', require('./api/subclasses'));
+router.use('/subraces', require('./api/subraces'));
+router.use('/traits', require('./api/traits'));
+router.use('/weapon-properties', require('./api/weaponProperties'));
 
 module.exports = router;
