+++ routes/api/abilityScores.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const AbilityScoreController = require('../controllers/api/abilityScoreController');
+const AbilityScoreController = require('../../controllers/api/abilityScoreController');
 
 router.get('/', AbilityScoreController.index);
 router.get('/:index', AbilityScoreController.show);
