+++ routes/api/conditions.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const ConditionController = require('../controllers/api/conditionController');
+const ConditionController = require('../../controllers/api/conditionController');
 
 router.get('/', ConditionController.index);
 router.get('/:index', ConditionController.show);
