+++ routes/api/features.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const FeatureController = require('../controllers/api/featureController');
+const FeatureController = require('../../controllers/api/featureController');
 
 router.get('/', FeatureController.index);
 router.get('/:index', FeatureController.show);
