+++ routes/api/languages.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const LanguageController = require('../controllers/api/languageController');
+const LanguageController = require('../../controllers/api/languageController');
 
 router.get('/', LanguageController.index);
 router.get('/:index', LanguageController.show);
