+++ routes/api/races.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const RaceController = require('../controllers/api/raceController');
+const RaceController = require('../../controllers/api/raceController');
 
 router.get('/', RaceController.index);
 router.get('/:index', RaceController.show);
