+++ routes/api/spellcasting.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const SpellcastingController = require('../controllers/api/spellcastingController');
+const SpellcastingController = require('../../controllers/api/spellcastingController');
 
 router.get('/', SpellcastingController.index);
 router.get('/:index', SpellcastingController.show);
