+++ routes/api/spells.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const SpellController = require('../controllers/api/spellController');
+const SpellController = require('../../controllers/api/spellController');
 
 router.get('/', SpellController.index);
 router.get('/:index', SpellController.show);
